requirements.txt
python-chess==1.999
numpy==1.24.3
tensorflow==2.13.0
pygame==2.5.2
scikit-learn==1.3.0
pandas==1.5.3
matplotlib==3.7.2
flask==2.3.3
flask-cors==4.0.0